feat(gazelle): A bazel-gazelle language extension for rules_cue
gazelle/cue contains a go_library which can be linked into a gazelle_binary to automate generation of bazel BUILD files for .cue files.

When a .cue file declares a package and the package name matches the directory in which it lives, it becomes part of a cue_library(name="cue_default_library", ...) target.

When a .cue file does not declare a package, it is turned into a cue_binary named after the file's base name, e.g. foo.cue becomes cue_binary(name="foo").