feat: Allow package names to not match parent directory name
According to the CUE documentation, A package is identified by its import path and its package name, separated by a :. If its name is equal to the base of this path it may be omitted.

We modify our gazelle extension to correctly declare the importpath for such packages (instead of ignoring them), and we modify the cue_library rule to strip off the package name from the importpath when building the pkg zip.

To exercise this new feature, we extend examples/hello_world to import examples/lang with English and German versions.